% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity_GERM.R
\name{GERM}
\alias{GERM}
\title{Generalized Riemannian Multidimensional Scaling}
\usage{
GERM(diss, p = 2, geometry = c("euclidean", "spherical", "hyperbolic"), ...)
}
\arguments{
\item{diss}{a S3 \code{\link[stats]{dist}} object of distance/dissimilarity among \eqn{n} objects.}

\item{p}{dimension of the target low-dimensional manifold (default: 2).}

\item{geometry}{name of the target geometry; one of \code{"euclidean"},\code{"spherical"}, and \code{"hyperbolic"} (default: \code{"euclidean"}).}
}
\value{
a named list containing \describe{
\item{embedding}{low-dimensional embedding of \eqn{n} objects. If \code{geometry="euclidean"}, it returns an \eqn{(n\times p)} matrix. Otherwise, it returns an \eqn{(n\times p+1)} matrix. In both cases, each row corresponds to an embedded vector onto the chosen manifold.}
}
}
\description{
Generalized Riemannian Multidimensional Scaling
}
\concept{dissimilarity}
