% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity_GERM.R
\name{germ}
\alias{germ}
\title{Generalized Riemannian Multidimensional Scaling}
\usage{
germ(diss, p = 2, geometry = c("euclidean", "spherical", "hyperbolic"), ...)
}
\arguments{
\item{diss}{a S3 \code{\link[stats]{dist}} object of distance/dissimilarity among \eqn{n} objects.}

\item{p}{dimensionality of the target low-dimensional manifold (default: 2).}

\item{geometry}{name of the target geometry; one of \code{"euclidean"},\code{"spherical"}, and \code{"hyperbolic"} (default: \code{"euclidean"}).}
}
\description{
Generalized Riemannian Multidimensional Scaling
}
\concept{dissimilarity}
